moved Type UserCache to global type def
