+++ src/lib/stores/UserCache.ts
@@ -9,12 +9,3 @@ export default writable({
     profile_picture: undefined,
     user_role: undefined,
 });
-
-export interface TypeUserCache {
-    _id: string | undefined
-    full_name: string | undefined
-    email: string | undefined
-    username: string | undefined
-    profile_picture: string | undefined
-    user_role: USER_ROLE | undefined
-}
